% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{FindDirection}
\alias{FindDirection}
\title{Find Direction of Smallest Weighted Projected data variance}
\usage{
FindDirection(data, x, w)
}
\arguments{
\item{data}{Data as matrix [d,n].}

\item{x}{Location as vector [d].}

\item{w}{Weights as vector [n].}
}
\value{
List of \describe{ \item{v}{Vector [d] which minimizes the
  variance of the weighted projected data.}
  \item{directional_variance}{weighted projected data variance
  in direction of \code{v}} }
}
\description{
Finding the direction minimizing the variance of the weighted projected data.
}
