% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{FindModeNewton}
\alias{FindModeNewton}
\title{Newton Method finding closest mode}
\usage{
FindModeNewton(y = NULL, lcd, h, reweighted = TRUE, delta = 1e-10,
  max_iter = 20)
}
\arguments{
\item{y}{Starting point (default: the mode of the non-smooth estimator)}

\item{lcd}{Fitted log-concave denstiy of class \code{dlc} from package
\code{logcondens}.}

\item{h}{Re-weighting bandwidth}

\item{delta}{accuracy in the Newton method.}
}
\value{
mode of the smooth (re-weigted) density.
}
\description{
A Newton method finding the closest mode to zero of the smooothed
(re-weighted) density.
}
