% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{FindReweightedMode}
\alias{FindReweightedMode}
\title{Find Reweighted Mode}
\usage{
FindReweightedMode(z, w, h)
}
\arguments{
\item{z}{Projected data [n]}

\item{w}{Weights of the data points [n]}

\item{h}{Bandwidth}
}
\value{
A list with the values: \item{m}{Mode}
  \item{logcon}{The fitted
  log-concave density of the function call
  \code{\link[logcondes]{activeSetLogCon}}}
  \item{reweighted_density}{The
  values of the reweighted density at the points of \code{logcon$x}}
}
\description{
Finding the univariate mode of the weighted projected data, which is
reweighted by the kernel.
}
