% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{FindSmoothedMode}
\alias{FindSmoothedMode}
\title{Find Smoothed Mode}
\usage{
FindSmoothedMode(x, phi, gam, isKnot = TRUE, F.smoothed = NULL,
  delta = 1e-10)
}
\arguments{
\item{x}{Vector [m] of support points of the data.}

\item{phi}{Vector [m] with the values of the log-concave density at the
support points.}

\item{gam}{Bandwidth parameter of the Gaussian smoothing.}

\item{isKnot}{Logical vector [m] with the information if the corresponding
support point is a knot of the log-concave density. By default all support
points are treated as knots.}

\item{F.smoothed}{Vector [m] with the values of the smoothe log-concave
density at the support points caluclated with bandwidth \code{gam}; see
also \pkg{logcondens}. If not \code{NULL} the starting interval of the
bisection search will be chosen accordingly.}

\item{delta}{Stopping criteria. (default = 1e-10)}
}
\value{
Mode of the smoothed log-concave density.
}
\description{
Finding the mode of the smoothed log-concave density with a bisection search.
}
