% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{GammaCircle}
\alias{GammaCircle}
\title{Gamma distributed on circle}
\usage{
GammaCircle(n = 200, shape = 5, rate = 1, seed = NULL)
}
\arguments{
\item{n}{Sample size}

\item{seed}{Change the seed for RNG}
}
\description{
Gamma distributed on circle
}
