% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{GammaLine}
\alias{GammaLine}
\title{Gamma on a line}
\usage{
GammaLine(n, shape = 2, rate = 4, seed = NULL)
}
\arguments{
\item{n}{Sample size}

\item{seed}{Change the seed for RNG}
}
\description{
2d-data with X1 uniformlly distributed on [-5, 5] and X2 gamma distributed
with certain shape and rate. Note that the true ridge is the line with X2 =
(shape - 1) / rate
}
