% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{GammaTripleLine}
\alias{GammaTripleLine}
\title{Gamma tiple line}
\usage{
GammaTripleLine(n, shape = 2, rate = 4, distance = c(0, 1, 4),
  seed = NULL)
}
\arguments{
\item{n}{Sample size}

\item{seed}{Change the seed for RNG}
}
\description{
The true ridgeset of the data is approximatly at X2 = 1/4, 5/4, 4 + 1/4.
}
