% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{geom_Arrows}
\alias{geom_Arrows}
\title{Arrow geom}
\usage{
geom_Arrows(tib, ...)
}
\arguments{
\item{tib}{A tibble containing X1_start, X2_start, X1, X2}
}
\description{
Plots an arrow between each starting and endpoint
}
