% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{geom_PostInt}
\alias{geom_PostInt}
\title{Plot post intervals}
\usage{
geom_PostInt(interval, ...)
}
\arguments{
\item{interval}{Tibble with the variables: X1_start,
X2_start, X1_end, X2_end}
}
\description{
Plot post intervals
}
