% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{geom_StepPath}
\alias{geom_StepPath}
\title{Plot step Path}
\usage{
geom_StepPath(steps, ...)
}
\arguments{
\item{steps}{List of steps generated by the "save" option in "LogConRidge".}
}
\description{
Plot step Path
}
