% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{geom_StepPoints}
\alias{geom_StepPoints}
\title{Plot Step Points}
\usage{
geom_StepPoints(steps, ...)
}
\arguments{
\item{steps}{List of steps generated by the "save" option in "LogConRidge".}
}
\description{
Plot Step Points
}
