% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{GetCI}
\alias{GetCI}
\title{Get Confidence Interval}
\usage{
GetCI(data, x, h, level = 0.05, alpha = 1)
}
\arguments{
\item{data}{Data as dataframe/tibble [d, n].}

\item{x}{Either a vector [d] or a tibble [m, d] with column-names \code{X1,
X2} corresponding to m different points}

\item{h}{Bandwidth (sould correspond to the one \code{x} was estimated with).}

\item{level}{The confidence level of each confidence interval is \code{1 -
level}.}

\item{alpha}{Cutoff value of the weight vector in percent.}
}
\value{
A named vector or a tibble with the columns \code{X1, X2, X1_left,
  X2_left, X1_right, X2_right}.
}
\description{
Calculates the confidence interval of the mode for the weighted projected
data in direction of minimal variance.
}
\details{
The points in \code{x} should correspond to the estimated ridgepoints of the
loc-concave method. Furthermore, \code{h} and \code{alpha} should be the
bandwidth which has been used for estimating \code{x}. The returned values
give the (\code{1 - level})-confidence interval for each point in \code{x}.

Whenever \code{x} is not a vector, a for-loop is used. To speed up things,
one may use vectors for \code{x} with some multi-threat method outside the
function.
}
