% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{GetGrid}
\alias{GetGrid}
\title{Get Grid}
\usage{
GetGrid(data, d, r = NULL)
}
\arguments{
\item{data}{data as matrix [d, n]}

\item{d}{distance between grid points}

\item{r}{maximal distance of each grid point from data. (default r = d)}
}
\value{
matrix [d, n] containing one grid point per column.
}
\description{
Calculates a starting point grid with some given maximal distance to the data.
}
