% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{GetMeanShift}
\alias{GetMeanShift}
\title{Mean Shift}
\usage{
GetMeanShift(data, x, h)
}
\arguments{
\item{data}{data as matrix [d, n].}

\item{x}{current point [d].}

\item{h}{Bandwidth of the Gaussian kernel}
}
\value{
A list with the following values:
\item{m}{Mean shift vector [d].}
\item{w}{Weights for each datapoint [n].}
\item{loc_data}{Localized data by \code{x} and \code{h} [d, n].}
}
\description{
Find the mean shift at the current point.
}
