% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{GetNewtonStep}
\alias{GetNewtonStep}
\title{Newton step for mode of smoothed (re-weighted) density}
\usage{
GetNewtonStep(y, lcd, h, reweighted = TRUE, knots, delta = 1e-10/2)
}
\arguments{
\item{y}{Starting point}

\item{lcd}{Fitted log-concave density of class \code{dlc} from package
\code{logcondens}.}

\item{h}{Re-weighting bandwidth}

\item{reweighted}{Which function to use. The re-weighted one or not.}

\item{knotes}{Knotes of the log-concave density estimator}
}
\description{
Newton step with step-size correction, if \code{f2 < 0}. Ohterwise, goes to
direction of the next knot.
}
