% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{GetPath}
\alias{GetPath}
\title{Get Path tibble}
\usage{
GetPath(steps)
}
\arguments{
\item{steps}{List of steps generated by the "save" option in "LogConRidge".}
}
\value{
Tibble with columns X1_start, X2_start, X1_end, X2_end and each row
corresponds to one step in the LCRS.
}
\description{
Get Path tibble
}
