% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{GetProjectedMeanShift}
\alias{GetProjectedMeanShift}
\title{Subsapce Projection}
\usage{
GetProjectedMeanShift(data, x, h)
}
\arguments{
\item{data}{data as matrix [d, n].}

\item{h}{Bandwidth}

\item{m}{vector to project [d].}
}
\value{
Projected vector [d].
}
\description{
Finds the subspace and projects the given vector on it
}
