% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{GetSmoothingBandwidth}
\alias{GetSmoothingBandwidth}
\title{Get Smoothing Bandwidth}
\usage{
GetSmoothingBandwidth(dlc)
}
\arguments{
\item{logcon}{Object of class `dlc` from package \code{link[logcondens]{LogConDens}}.}
}
\value{
Smoothing Bandwidth
}
\description{
Returns the bandwith for the smoothed density, such that that its variance
co-incides with the variance of the empirical distribution.
}
