% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{GetTibble}
\alias{GetTibble}
\title{Create Tibble}
\usage{
GetTibble(grid, result, h, n, method, data_name, seed, alpha)
}
\arguments{
\item{grid}{Matrix of the starting points [d, n]}

\item{result}{Matrix of the limits [d, n]}

\item{h}{Used Bandwith}

\item{method}{Used method}

\item{data_name}{Name of the used data generation}

\item{seed}{Used seed for the data generation}

\item{alpha}{Used alpha for weight vector}
}
\description{
Creates a Tibble of the estimated ridge points
}
\details{
The Tibble has the following columns:
X1_start, X2_start, X1, X2, h, n, method, data, seed, date, time
}
