% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{LogConRidge}
\alias{LogConRidge}
\title{Log-Concave Ridge search}
\usage{
LogConRidge(data, x, h, delta = 1e-06, print = FALSE, save = FALSE,
  max_iter = 20, reweight = TRUE, smooth = FALSE, alpha = 1)
}
\arguments{
\item{data}{Data as  matrix [d, n].}

\item{x}{Starting location as vector [d].}

\item{h}{Bandwidth for the weight function}

\item{delta}{Stopping criteria (default: \code{delta = 1e-6}).}

\item{print}{If TRUE prints the values at each step (default \code{print =
FALSE}).}

\item{save}{If TRUE returns a list, which contains for each step a list with
startpoint, endpoint, direction, log-concave density.}

\item{max_iter}{The maximal amount of iterations. If the maximal iterations
are reached a warning will be printed.}

\item{alpha}{Cutoff value for the weight vector in percent.}

\item{smoothed}{Logical, if \code{TRUE} the mode of the smoothed log-concave
density estimator will be used. (default: \code{FALSE})}
}
\value{
Vector [d] the algorithm converges to on the ridge.
}
\description{
Finds the closest Ridge Point using log-concave variante.
}
\examples{
data <- SimCircle(200)
x <- c(-0.5, 0.1)
h <- 0.5
LogConRidge(data, x, h, print = TRUE)
}
