% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{NewWeights}
\alias{NewWeights}
\title{Weight cutoff}
\usage{
NewWeights(w, alpha = 0.99)
}
\arguments{
\item{w}{Vector of weights [n]}

\item{alpha}{Percentage of weighted data to keep}
}
\value{
A list with the values:
\item{w}{Vector of new weights [n]}
\item{isPositive}{Logical vector wheter or not the entry in \code{w} is 0 or not. [n]}
}
\description{
Cutoff weights with very small influance
}
