% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{Plot1Step}
\alias{Plot1Step}
\title{Plots the log-concave density of a given step}
\usage{
Plot1Step(list_element)
}
\arguments{
\item{list_element}{Contains one list element generated by the "save"
option in "LogConRidge".}
}
\description{
Plots the log-concave density of a given step
}
