% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{PlotArrows}
\alias{PlotArrows}
\title{Arrow Plot}
\usage{
PlotArrows(grid, ridge)
}
\arguments{
\item{grid}{Matrix of starting points [d, n].}

\item{ridge}{Matrix of end points [d, n].}
}
\description{
Plots an arrow between each starting and endpoint
}
