% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{PlotDensity}
\alias{PlotDensity}
\title{Weighted Density Plot}
\usage{
PlotDensity(z, w, h = 1, which = c(1, 5))
}
\arguments{
\item{z}{Univariate data [n].}

\item{w}{Weights of the data [n].}

\item{h}{Bandwith needed for the smoothed densities. (Default = 1)}

\item{which}{A (sub)vector of \code{1:5}. Each number
corresponding to a certain estimator: 1: log-concave, 2: smoothed
log-concave, 3: reweigted log-concave, 4: reweigted smoothed log-concave, 5: KDE.}
}
\value{
ggplot2 object showing the desired densities.
}
\description{
Plots of the estimated densities for different methods.
}
