% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{PlotDensityBoth}
\alias{PlotDensityBoth}
\title{Weighted and Log-Concave Density Plot}
\usage{
PlotDensityBoth(z, w, x = NULL, v = NULL)
}
\arguments{
\item{z}{Projected Data [n].}

\item{w}{Weights of the projected data [n].}

\item{x}{Location vector [d]. If \code{x} and \code{v} are \code{NULL} the
information will be plotted in the subtitle.}

\item{v}{Direction vector [d]. If \code{x} and \code{v} are \code{NULL} the
information will be plotted in the subtitle.}
}
\description{
Weighted and Log-Concave Density Plot
}
