% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{PlotReweightedDensity}
\alias{PlotReweightedDensity}
\title{Plot reweighted Density}
\usage{
PlotReweightedDensity(z, w, h)
}
\arguments{
\item{z}{Projected data [n]}

\item{w}{Weights [n]}

\item{h}{Bandwidth}
}
\description{
Plots the univariate log-concave density reweighted with the Gaussian kernel.
}
