% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{PlotSmoothedDensity}
\alias{PlotSmoothedDensity}
\title{Plot smoothed density}
\usage{
PlotSmoothedDensity(z, w)
}
\arguments{
\item{z}{Projected data [n]}

\item{w}{Weights [n]}

\item{gam}{Smoothing bandwidth}
}
\description{
Plot smoothed density
}
