% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{PostInterval}
\alias{PostInterval}
\title{Post Interval}
\usage{
PostInterval(data, x, h, thres = 0.9, alpha = 1, print = FALSE)
}
\arguments{
\item{x}{Matrix of the points to consider [d, n]}

\item{thres}{percent value according to the maximum of the density}
}
\description{
Calculates for each ridge point calculated by "lc" an interval, where
the mode could also lie, based on a threshold.
}
