% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{PostSmoothing}
\alias{PostSmoothing}
\title{Post Smoothing}
\usage{
PostSmoothing(data, x, h, alpha = 1)
}
\arguments{
\item{x}{Matrix of the points to consider [d, n]}
}
\description{
After calculating the "lc"-algorithm, it performce a mode smoothing after
convergence.
}
