% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{ProjectData}
\alias{ProjectData}
\title{Data projectction}
\usage{
ProjectData(data, x, v)
}
\arguments{
\item{data}{Data points given as a matrix.}

\item{x}{Location as a vector.}

\item{v}{Direction as a vector.}
}
\value{
The transformed and projected points. Either as a vector or dataframe
  if multiple points or directions are considered.
}
\description{
Transforms the data by \eqn{- \boldsymbol{x}} and projects it on the
direction \eqn{\boldsymbol{v}}
}
\examples{
data <- SimCirlce(200)
x    <- c(-1, 0)
v    <- c( 1, 0)
z <- ProjectData(data, x, v)
ggplot(dataframe(z = t(z)), aes(z)) + geom_density()
}
