% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{SaveTibble}
\alias{SaveTibble}
\title{Save tibble}
\usage{
SaveTibble(tib, name, add = TRUE)
}
\arguments{
\item{name}{Name of the file}

\item{add}{If \code{TRUE}, the data is added to the file, oterhwise the file
is overwritten.}

\item{tibble}{Tibble to save}
}
\description{
Save the tibble under given name
}
