% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{SCMS}
\alias{SCMS}
\title{Subspace Constraint Mean Shift}
\usage{
SCMS(data, x, h, delta = 1e-06, print = FALSE, iter_max = 100)
}
\arguments{
\item{data}{Data as matrix [d, n]}

\item{x}{Starting point [d]}

\item{h}{Bandwidth of the kernel density estimator}

\item{delta}{Achivent accuracy}

\item{print}{If \code{TRUE}, the path of the algorithm is plotted.}
}
\value{
Point on the ridge closest to \code{x} as vector
}
\description{
Implementation of the subspace constraint mean shift algorithm
}
