% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{SimSinus}
\alias{SimSinus}
\title{Simulate Sinus data}
\usage{
SimSinus(n, int = c(-2 * pi, 2 * pi), sd = 0.2, seed = NULL)
}
\arguments{
\item{n}{Sample size}

\item{sd}{Standard deviation}

\item{seed}{Change the seed for RNG}
}
\value{
Matrix with rows \code{X1, X2}. [2, n]
}
\description{
Generates data distributed along a sinus line with Gaussian noise having
standard deviation \code{sd}.
}
