% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{Weighted_Variance2}
\alias{Weighted_Variance2}
\title{Weighted Variance}
\usage{
Weighted_Variance2(z, w)
}
\arguments{
\item{z}{Points with positive probability as vector.}

\item{w}{Probability weights as vector.}
}
\value{
Variance of the finite distribution
}
\description{
Calculates the variance of a finite distribution with probability weigth \code{w} at \code{z}
}
