% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{Weights}
\alias{Weights}
\title{Weights}
\usage{
Weights(data, x, h, type = "Gaussian")
}
\arguments{
\item{data}{Data in a matrix.}

\item{x}{Location as a vector.}

\item{h}{Bandwidth parameter.}

\item{type}{Type of the used kernel function.}
}
\value{
Vector of the probability weights corresponding to the data.
}
\description{
Calculats the weights of the data for a given point \code{x}
}
